Refactor Topics from Homepage into its own file
